+++ Admin/src/main/java/com/example/admin/Entity/ImportGoodsEntity.java
@@ -20,5 +20,5 @@ public class ImportGoodsEntity {
 
     private int quantity;
 
-    private Long totalPrice;
+    private double totalPrice;
 }
